+++ saleor/graphql/webhook/subscription_types.py
@@ -1187,6 +1187,37 @@ class FulfillmentCreated(SubscriptionObjectType, FulfillmentBase):
         return data["notify_customer"]
 
 
+class FulfillmentUpdated(SubscriptionObjectType, FulfillmentBase):
+    notify_customer = graphene.Boolean(
+        description=(
+            "If true, send an email notification to the customer." + ADDED_IN_315
+        ),
+        required=True,
+    )
+
+    class Meta:
+        doc_category = DOC_CATEGORY_ORDERS
+        root_type = None
+        enable_dry_run = False
+        interfaces = (Event,)
+        description = "Event sent when new fulfillment is updated." + ADDED_IN_315
+
+    @staticmethod
+    def resolve_fulfillment(root, info: ResolveInfo):
+        _, data = root
+        return data["fulfillment"]
+
+    @staticmethod
+    def resolve_order(root, info: ResolveInfo):
+        _, data = root
+        return data["order"]
+
+    @staticmethod
+    def resolve_notify_customer(root, _info: ResolveInfo):
+        _, data = root
+        return data["notify_customer"]
+
+
 class FulfillmentCanceled(SubscriptionObjectType, FulfillmentBase):
     class Meta:
         root_type = "Fulfillment"
@@ -2282,6 +2313,7 @@ WEBHOOK_TYPES_MAP = {
     WebhookEventAsyncType.INVOICE_DELETED: InvoiceDeleted,
     WebhookEventAsyncType.INVOICE_SENT: InvoiceSent,
     WebhookEventAsyncType.FULFILLMENT_CREATED: FulfillmentCreated,
+    WebhookEventAsyncType.FULFILLMENT_UPDATED: FulfillmentUpdated,
     WebhookEventAsyncType.FULFILLMENT_CANCELED: FulfillmentCanceled,
     WebhookEventAsyncType.FULFILLMENT_APPROVED: FulfillmentApproved,
     WebhookEventAsyncType.FULFILLMENT_METADATA_UPDATED: FulfillmentMetadataUpdated,