+++ CHANGELOG.md
@@ -107,6 +107,8 @@ Shipping methods can be removed by the user after it has been assigned to a chec
   - Called after `requestPasswordReset` or `customerCreate` mutation for staff users.
 - Add `NOTIFY_CUSTOMER` flag to `FulfillmentCreated` type - #13620, by @Air-t
   - Informs apps if customer should be notified when fulfillment is created.
+- Add `FULFILLMENT_UPDATED` webhook event - #13630, by @Air-t
+  - Informs apps if customer should be notified when fulfillment is updated.
 
 ### Other changes
 - Add possibility to log without confirming email - #13059 by @kadewu